+++ django/applications/catmaid/static/js/WindowMaker.js
@@ -854,7 +854,13 @@ var WindowMaker = new function()
               WA.options.skeleton_node_scaling = Math.max(0, this.value) || 1.0;
               WA.adjustContent();
               WA.updateSkeletonNodeHandleScaling(this.value);
-        }, 5);
+        }, 3);
+
+    var textScalingInput = DOM.appendNumericField(tabs['View settings'],
+        'Text scaling', 'Scaling of text.', o.text_scaling, null, function() {
+              let value = parseInt(this.value, 10);
+              WA.updateTextScaling(value);
+            }, 3);
 
     DOM.appendToTab(tabs['Stacks'],
         [